Refactor: Move Campground Routes to Separate File
Moved all campground routes to their own file using the Router object.
