+++ app.js
@@ -11,6 +11,8 @@ const Review = require("./models/review");
 const methodOverride = require("method-override");
 const ejsMate = require("ejs-mate");
 
+const campgroundsRoute = require('./routes/campgrounds')
+
 // use ejs-locals for all ejs templates:
 app.engine("ejs", ejsMate);
 
@@ -34,19 +36,19 @@ async function main() {
 }
 
 // Middleware to Validate a new campground using JOI
-const validateCampground = (req, res, next) => {
-  // use the JOI .validate method to get any errors
-  // take the 'error' key from the resulting object after .validate(req.body)
-  const { error } = campgroundSchema.validate(req.body);
-  if (error) {
-    // use map to iterate over each object in the 'details' array and extract the error message
-    const msg = error.details.map((element) => element.message).join(",");
-    // console.log(msg);
-    throw new ExpressError(msg, 400);
-  } else {
-    next();
-  }
-};
+// const validateCampground = (req, res, next) => {
+//   // use the JOI .validate method to get any errors
+//   // take the 'error' key from the resulting object after .validate(req.body)
+//   const { error } = campgroundSchema.validate(req.body);
+//   if (error) {
+//     // use map to iterate over each object in the 'details' array and extract the error message
+//     const msg = error.details.map((element) => element.message).join(",");
+//     // console.log(msg);
+//     throw new ExpressError(msg, 400);
+//   } else {
+//     next();
+//   }
+// };
 
 const validateReview = (req, res, next) => {
   const { error } = reviewSchema.validate(req.body);
@@ -59,69 +61,12 @@ const validateReview = (req, res, next) => {
   }
 }
 
+app.use('/campgrounds', campgroundsRoute)
+
 // ROUTES
 app.get("/", (req, res) => {
   res.render("home");
 });
-
-app.get("/campgrounds", catchAsync(async (req, res) => {
-  const allCampgrounds = await Campground.find({});
-  res.render(`campgrounds/index`, { allCampgrounds });
-})
-);
-
-app.get("/campgrounds/new", (req, res) => {
-  res.render("campgrounds/newCampground");
-});
-
-app.post("/campgrounds", validateCampground, catchAsync(async (req, res, next) => {
-  // if (!req.body.campground) {
-  // 	throw new ExpressError("Invalid Campground Data", 400);
-  // }
-
-  const campground = new Campground(req.body.campground);
-  await campground.save();
-  res.redirect(`/campgrounds/${campground._id}`);
-})
-);
-
-app.get("/campgrounds/:id/edit", catchAsync(async (req, res) => {
-  const { id } = req.params;
-  const campground = await Campground.findById(id);
-  res.render("campgrounds/editCampground", { campground });
-})
-);
-
-app.put("/campgrounds/:id", validateCampground, catchAsync(async (req, res) => {
-  const { id } = req.params;
-  const updatedCampground = await Campground.findByIdAndUpdate(
-    id,
-    { ...req.body.campground },
-    {
-      runValidators: true,
-      new: true,
-    }
-  );
-
-  console.log(updatedCampground);
-  res.redirect(`/campgrounds/${updatedCampground._id}`);
-})
-);
-
-app.delete("/campgrounds/:id", catchAsync(async (req, res) => {
-  const { id } = req.params;
-  const deleteCampground = await Campground.findByIdAndDelete(id);
-  res.redirect("/campgrounds");
-})
-);
-
-app.get("/campgrounds/:id", catchAsync(async (req, res) => {
-  const { id } = req.params;
-  const campground = await Campground.findById(id).populate("reviews");
-  // console.log(campground);
-  res.render("campgrounds/showDetails", { campground });
-})
-);
 
 // allow user to delete a review
 app.delete(
@@ -141,20 +86,23 @@ app.delete(
   })
 );
 
-app.post("/campgrounds/:id/reviews", validateReview, catchAsync(async (req, res) => {
-  console.log(req.params);
-  const { id } = req.params;
-  const campground = await Campground.findById(id);
+app.post(
+  "/campgrounds/:id/reviews",
+  validateReview,
+  catchAsync(async (req, res) => {
+    console.log(req.params);
+    const { id } = req.params;
+    const campground = await Campground.findById(id);
 
-  const review = new Review(req.body.review);
+    const review = new Review(req.body.review);
 
-  campground.reviews.push(review);
+    campground.reviews.push(review);
 
-  await review.save();
-  await campground.save();
+    await review.save();
+    await campground.save();
 
-  res.redirect(`/campgrounds/${campground._id}`);
-})
+    res.redirect(`/campgrounds/${campground._id}`);
+  })
 );
 
 // Catch all for routes that don't match above routes
